+++ tests/models/test_snapshot.py
@@ -37,3 +37,18 @@ def test_load_from_json():
     assert snapshot.min_disk_size == data['min_disk_size']
     assert snapshot.size_gigabytes == data['size_gigabytes']
     assert snapshot.tags == data['tags']
+
+
+def snapshot_model_matches(snapshot, expected):
+    """
+    Helper function to check the Snapshot model against the expected dict.
+    """
+    return snapshot.id == expected['id'] and \
+           snapshot.name == expected['name'] and \
+           snapshot.regions == expected['regions'] and \
+           snapshot.created_at == expected['created_at'] and \
+           snapshot.resource_id == expected['resource_id'] and \
+           snapshot.resource_type == expected['resource_type'] and \
+           snapshot.min_disk_size == expected['min_disk_size'] and \
+           snapshot.size_gigabytes == expected['size_gigabytes'] and \
+           snapshot.tags == expected['tags']
